+++ scout/models/case/case.py
@@ -22,6 +22,7 @@ individual = dict(
     mother=str,  # Individual id of mother
     capture_kits=list,  # List of names of capture kits
     bam_file=str,  # Path to bam file
+    vcf2cytosure=str, # Path to CGH file
     analysis_type=str,  # choices=ANALYSIS_TYPES
     confirmed_sex=bool,  # True or False. None if no check has been done
     confirmed_parent=bool,
@@ -79,6 +80,7 @@ case = dict(
     diagnosis_genes=list,  # List of references to genes
 
     has_svvariants=bool,  # default=False
+    has_strvariants=bool, # default=False
 
     is_migrated = bool, # default=False