+++ tests/load/test_load_panel.py
@@ -1,3 +1,5 @@
+from pprint import pprint as pp
+from scout.load.panel import load_panel
 
 def test_load_panel(gene_database, panel_info):
     # GIVEN an database with genes but no panels
@@ -5,11 +7,11 @@ def test_load_panel(gene_database, panel_info):
     assert adapter.gene_panels().count() == 0
     
     # WHEN loading a gene panel
-    adapter.load_panel(
-        path=panel_info['file'], 
+    load_panel(
+        panel_path=panel_info['file'], 
+        adapter=adapter,
         institute=panel_info['institute'], 
         panel_id=panel_info['panel_name'], 
-        date=panel_info['date'], 
         panel_type=panel_info['type'], 
         version=panel_info['version'], 
         display_name=panel_info['full_name']