+++ tests/parse/test_parse_compounds.py
@@ -4,7 +4,7 @@ from scout.parse.variant.compound import parse_compounds
 
 def test_parse_simple_compound(case_obj):
     ## GIVEN a compound string
-    compound_string = "643594:7_117175579_AT_A>32"
+    compound_string = "internal_id:7_117175579_AT_A>32"
 
     ## When parsing the compounds
     compounds = parse_compounds(compound_string, case_id=case_obj['_id'], variant_type='clinical')
@@ -16,7 +16,7 @@ def test_parse_simple_compound(case_obj):
 
 def test_parse_compound_no_score(case_obj):
     ## GIVEN a compound string
-    compound_string = "643594:7_117175579_AT_A"
+    compound_string = "internal_id:7_117175579_AT_A"
 
     ## When parsing the compounds
     compounds = parse_compounds(compound_string, case_id=case_obj['_id'], variant_type='clinical')
@@ -38,11 +38,10 @@ def test_parse_compound_no_compound(case_obj):
 
 def test_parse_multiple_compound(case_obj):
     ## GIVEN a compound string
-    compound_string = "643594:7_117175579_AT_A>32|7_117175580_T_G>28"
+    compound_string = "internal_id:7_117175579_AT_A>32|7_117175580_T_G>28"
 
     ## When parsing the compounds
     compounds = parse_compounds(compound_string, case_id=case_obj['_id'], variant_type='clinical')
 
     ## THEN assert that the correct info is returned
     assert len(compounds) == 2
-